call jena from util, not from service
